Begin porting the securesd script to C
We need to port the script to C so that we can install
it as a suid binary. That way root permissions are not
necessary to mount, so we don't need to hack the init
scripts to mount at boot time. Once this binary is
installed the Android app will be able to mount/unmount
without root permission.

It'll be nice if we can get cryptsetup compiled and merge
this functionality into one binary.