Auto-mount moved to Android (Java) code
The automatic mount/unmount code has been moved to
the java code. Now there's no need to hack the init
scripts. The app can now also mount and unmount without
root permission as long as the securesd binary is
properly installed.

Note: On most devices loop device nodes are not
created at boot, so an init script is still necessary
to create loop devices as needed. The next step is to
create the device nodes as needed so that no init
script hacking is needed at all.